+++ src/plugins/danielo515/tiddlypouch/startup-pouch.js
@@ -21,11 +21,11 @@ The existence of the database determines if the plugin will be active or not.
     exports.name = "pouchdb";
     exports.before = ["startup"];
     exports.platforms = ["browser"];
-    exports.synchronous = true;
+    exports.synchronous = false;
 
     var CONFIG_PREFIX = "$:/plugins/danielo515/tiddlypouch/config/";
 
-    exports.startup = function () {
+    exports.startup = function (callback) {
         /* --- Declaration ZONE ---*/
         //============================
 
@@ -59,6 +59,14 @@ The existence of the database determines if the plugin will be active or not.
             return ddoc;
         }
 
+        /**
+         * Creates generic conflict-handler functions.
+         * The returned function logs a default message to the console in case of conflict,
+         * otherwise it throws the error so the next catch on the promise chain can handle it 
+         * 
+         * @param {any} message the message the returned handler will log to the console in case of conflict
+         * @returns {function} handler a function ready to be used inside a catch statement in a promise chain
+         */
         function conflict (message){
             return function(err){
                 if (err.status == 409) {
@@ -67,7 +75,8 @@ The existence of the database determines if the plugin will be active or not.
                 throw err;
             }
         }
-
+        
+        /** The indexes that are going to be created on the local database */
         var indexes = 
         { 
             by_type: createDesignDoc('by_type', 
@@ -101,16 +110,21 @@ The existence of the database determines if the plugin will be active or not.
             logger.log("Index by type created");
         }).catch(
             conflict("Index by_type exists already ")
-        );
-
+        )
         /*Fetch and add the StoryList before core tries to save it*/
-        $tw.TiddlyPouch.database.get("$:/StoryList").then(function (doc) {
+        .then(function() {
+            return $tw.TiddlyPouch.database.get("$:/StoryList")
+        }).then(function (doc) {
             $tw.wiki.addTiddler(new $tw.Tiddler(doc.fields, { title: doc._id, revision: doc._rev }));
-            logger.log("StoryList is already in database ", doc.fields);
+            logger.log("StoryList was already in database ", doc.fields);
+            return $tw.TiddlyPouch.database.get("$:/DefaultTiddlers")
+        }).then(function(doc) {
+           $tw.wiki.addTiddler(new $tw.Tiddler(doc.fields, { title: doc._id, revision: doc._rev }));
+            logger.log("Default tiddlers loaded from database ", doc.fields);
         }).catch(function (err) {
-            logger.log("Error retrieving StoryList");
+            logger.log("Error retrieving StoryList or DefaultTiddlers");
             logger.log(err);
-        });
+        }).then(callback);
 
     };